add TextSplitter base class
