+++ chunkers/character_chunker.py
@@ -1,8 +1,13 @@
-from typing import Any, List, Literal, Optional, Union, Callable, Iterable
+from typing import Any, List, Literal, Optional, Union, Callable
 import re
-import warnings
 
-class RecursiveCharacterTextSplitter:
+try:
+    from ..chunkers.base_chunker import TextSplitter
+except:
+    from chunkers.base_chunker import TextSplitter
+
+
+class RecursiveCharacterTextSplitter(TextSplitter):
     """Splitting text by recursively look at characters.
 
     Recursively tries to split by different characters to find one
@@ -11,30 +16,18 @@ class RecursiveCharacterTextSplitter:
     Adapted from Langchain:
     https://github.com/langchain-ai/langchain/blob/0606aabfa39acb2ec575ea8bbfa4c8e662a6134f/libs/text-splitters/langchain_text_splitters/character.py#L58
     """
-    def __init__(
-        self,
-        chunk_size: int = 4000,
-        chunk_overlap: int = 200,
-        length_function: Callable[[str], int] = len,
-        add_start_index: bool = False,
-        strip_whitespace: bool = True,
-        separators: Optional[List[str]] = None,
-        keep_separator: Union[bool, Literal["start", "end"]] = True,
-        is_separator_regex: bool = False,
-        **kwargs: Any,
-    ) -> None:
+    def __init__(self, chunk_size: int = 4000, chunk_overlap: int = 200, length_function: Callable[[str], int] = len,
+                 add_start_index: bool = False, strip_whitespace: bool = True, separators: Optional[List[str]] = None,
+                 keep_separator: Union[bool, Literal["start", "end"]] = True, is_separator_regex: bool = False,
+                 **kwargs: Any) -> None:
         """Create a new TextSplitter."""
+        super().__init__(chunk_size, chunk_overlap, length_function, keep_separator, add_start_index, strip_whitespace)
         if chunk_overlap > chunk_size:
             raise ValueError(
                 f"Got a larger chunk overlap ({chunk_overlap}) than chunk size "
                 f"({chunk_size}), should be smaller."
             )
-        self._chunk_size = chunk_size
-        self._chunk_overlap = chunk_overlap
-        self._length_function = length_function
-        self._keep_separator = keep_separator
-        self._add_start_index = add_start_index
-        self._strip_whitespace = strip_whitespace
+
         self._separators = separators or ["\n\n", "\n", " ", ""]
         self._is_separator_regex = is_separator_regex
 
@@ -80,57 +73,6 @@ class RecursiveCharacterTextSplitter:
 
     def split_text(self, text: str) -> List[str]:
         return self._split_text(text, self._separators)
-
-    def _merge_splits(self, splits: Iterable[str], separator: str) -> List[str]:
-        # We now want to combine these smaller pieces into medium size
-        # chunks to send to the LLM.
-        separator_len = self._length_function(separator)
-
-        docs = []
-        current_doc: List[str] = []
-        total = 0
-        for d in splits:
-            _len = self._length_function(d)
-            if (
-                    total + _len + (separator_len if len(current_doc) > 0 else 0)
-                    > self._chunk_size
-            ):
-                if total > self._chunk_size:
-                    warnings.warn(
-                        f"Created a chunk of size {total}, "
-                        f"which is longer than the specified {self._chunk_size}"
-                    )
-                if len(current_doc) > 0:
-                    doc = self._join_docs(current_doc, separator)
-                    if doc is not None:
-                        docs.append(doc)
-                    # Keep on popping if:
-                    # - we have a larger chunk than in the chunk overlap
-                    # - or if we still have any chunks and the length is long
-                    while total > self._chunk_overlap or (
-                            total + _len + (separator_len if len(current_doc) > 0 else 0)
-                            > self._chunk_size
-                            and total > 0
-                    ):
-                        total -= self._length_function(current_doc[0]) + (
-                            separator_len if len(current_doc) > 1 else 0
-                        )
-                        current_doc = current_doc[1:]
-            current_doc.append(d)
-            total += _len + (separator_len if len(current_doc) > 1 else 0)
-        doc = self._join_docs(current_doc, separator)
-        if doc is not None:
-            docs.append(doc)
-        return docs
-
-    def _join_docs(self, docs: List[str], separator: str) -> Optional[str]:
-        text = separator.join(docs)
-        if self._strip_whitespace:
-            text = text.strip()
-        if text == "":
-            return None
-        else:
-            return text
 
 
 def _split_text_with_regex(
